:recycle: Add controller layer to task routes
